+++ chat/.gitignore
@@ -2,3 +2,4 @@ env/
 venv/
 *.pyc
 __pycache__/
+chat/__pycache__/
